removed the message "the list is empty"
The reason for this change is that the movies array will initially be empty, and it will remain empty while the fetching process takes place in the background. Until the data is finally fetched, you don't want to display the message “The list is empty” since you don’t yet know whether the fetched array will be empty or not.